refactor: context와 hook 분리
