+++ client/src/lib/hooks/useContextHooks.ts
@@ -2,6 +2,7 @@ import React, { useContext } from 'react';
 
 import { EpicStateContext, EpicDispatchContext } from '@/contexts/epicContext';
 import { StoryStateContext, StoryDispatchContext } from '@/contexts/storyContext';
+import { UserContext } from '@/contexts/userContext';
 
 export function useEpicState() {
   const state = useContext(EpicStateContext);
@@ -26,3 +27,15 @@ export function useStoryDispatch() {
   if (!dispatch) throw new Error('Cannot find StoryProvider'); // 유효하지 않을땐 에러를 발생
   return dispatch;
 }
+
+export const useUserState = () => {
+  const { userState } = useContext(UserContext);
+  if (!userState) throw new Error('Cannot find UserProvider');
+  return userState;
+};
+
+export const useUserDispatch = () => {
+  const { dispatch } = useContext(UserContext);
+  if (!dispatch) throw new Error('Cannot find UserProvider');
+  return dispatch;
+};
